Made malleable to other machines
Instead of referencing my specific path throughout the code, I made the program pull the cwd every time it was needed.
In addition, I added a chunk that would confirm if there was a "Lists" folder in the cwd, and in the case that it wasn't, it would make one for the sake of the code.